% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_datasheet.R
\name{analysis_cols}
\alias{analysis_cols}
\title{Create analysis dataframe}
\usage{
analysis_cols(x, method = NULL)
}
\arguments{
\item{x}{A data sheet containing formatted field metadata}

\item{method}{One of "BRUV" or "DOV"}
}
\description{
Formats standardised field metadata to output a data frame containing the columns
required for video analysis. Required for `make_datasheet()`.
}
