% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ilab_structure.R
\name{check_structure}
\alias{check_structure}
\title{Validate Campaign Directories Against a Template}
\usage{
check_structure(MP = NULL, campaign = NULL, root = get_dir("ilab_fish"))
}
\arguments{
\item{MP}{(Optional) acronym(s) of target marine park(s).}

\item{campaign}{(Optional) name of campaign(s) in the target marine park(s). 
Requires `MP` to be defined.}

\item{root}{A character path to the directory whose structure should be
validated (Default = iLab::get_dir("ilab_fish")).}
}
\value{
A list containing a summary data.frame detailing the matches per rule 
in the template specification, a data.frame of unexpected file paths found, and 
a data frame detailing all matches found.
}
\description{
`check_structure()` validates iLab-fish campaign directories against a template
specification (created by the internal function `campaign_structure`) that 
describes the allowed directory and file layout.
}
\examples{
\dontrun{

# Run the check on all of iLab-Fish
# check_structure()

# Run the check for specific campaign
checks <- check_structure(MP = "RIMR", campaign = "2023-02_BRUV")
# Print summary
checks$summary

# Unexpected files
checks$unexpected

# View matches
View(checks$matches)

}

}
