% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_metadata_functions.R
\name{clean_field_names}
\alias{clean_field_names}
\title{Clean field metadata variable names}
\usage{
clean_field_names(x)
}
\arguments{
\item{x}{A data frame containing field metadata.}
}
\description{
Standardises field metadata column names
}
\examples{
\dontrun{
field.data <- field.data\%>\%
  janitor::clean_names()\%>\%         # standardise separators and case
  clean_names_method(method = method) #updates synonyms depending on method
}

}
