% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_hds.R
\name{compare_hds}
\alias{compare_hds}
\title{Check for file differences between two hard-drives}
\usage{
compare_hds(raw_hd, bu_hd)
}
\arguments{
\item{raw_hd}{Directory for the Raw Hard drive (original files)}

\item{bu_hd}{Directory for the Backup Hard drive (copied files)}
}
\value{
A set of console messages highlighting errors/warnings.
}
\description{
`compare_hds()` checks for differences in files between two hard drives.
It looks for differing numbers of files, different file naming, different file sizes,
different modification dates, and runs preliminary checks for corrupt files.
}
\examples{
\dontrun{

# Run the check on all of iLab-Fish
# compare_hds(raw_hd = "E:/", bu_hd = "D:/)

}

}
