% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_metadata_functions.R
\name{field_variables}
\alias{field_variables}
\title{Get required field metadata columns}
\usage{
field_variables(method = NULL)
}
\arguments{
\item{method}{Character; one of `c("BRUV","DOV", "ROV")`. Determines the sheet
structure, required columns, and data validations applied.}
}
\description{
Get vectors of expected and vital variables required in field metadata dependent
on sampling method. Returns a list containing the named vectors `expected` and 
`vital`.
}
\examples{
\dontrun{
var <- fields_variables("BRUV")
}

}
