% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_datasheet_stashed.R
\name{make_absolute_stashed}
\alias{make_absolute_stashed}
\title{Convert excel range to absolute range}
\usage{
make_absolute_stashed(range)
}
\arguments{
\item{range}{A character vector containing Excel-style cell references
or ranges (e.g., `"A1"`, `"A1:B5"`, `"Sheet1!A1:B1"`, `c("A1:A5","C1:C5")`).}
}
\description{
Converts an excel range (e.g. A1:A5) into an absolute range (e.g. $A$1:$A$5)
This function operates using a simple regular expression and works for typical
cell references, including ranges and multi-area references.
}
