% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_campaign_dir.R
\name{make_campaign_dir}
\alias{make_campaign_dir}
\title{Make Campaign Directory}
\usage{
make_campaign_dir(
  marine.park = NULL,
  date = NULL,
  method = NULL,
  base.dir = get_dir("ilab_fish")
)
}
\arguments{
\item{marine.park}{character vector giving the marine park identifier. Defaults 
to \code{NULL} which prompts the user using a pop-up.}

\item{date}{character vector giving the campaigns year and month (e.g. 2023-01). 
Defaults to \code{NULL} which prompts the user using a pop-up.}

\item{method}{character vector giving the sample collection method. Must be one 
of 'BRUV', 'DOV', 'ROV', or 'NULL'. Defaults to \code{NULL} which prompts the 
user using a pop-up.}

\item{base.dir}{character vector setting the root directory where folders are 
created. Defaults to ilab-fish and should only be changed for testing.}
}
\description{
Interactive creation of new campaign directory following standardised templates,
or addition of missing folders to existing campaign directory folders. If any of 
\code{marine.park, date, or method == NULL} interactive prompts allow the user 
to set these details.
}
\examples{
\dontrun{
make_campaign_dir() # Interactive
make_campaign_dir(marine.park = "RIMR", date = "2025-02", method = "BRUV") # Provide file path
}

}
