% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_library.R
\name{ref_library}
\alias{ref_library}
\title{Shiny application displaying iLab-Fish reference library images}
\usage{
ref_library()
}
\value{
A **Shiny Application**
}
\description{
Starts a the iLab Fish reference library application, which allows users to
filter and display PNG/JPEG images stored in Asset Fish where the ParkID , 
genus, and species can be extracted from the filepath.
}
