% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton_EMOb.R
\name{skeleton_emob}
\alias{skeleton_emob}
\title{Create an EMOb Skeleton}
\usage{
skeleton_emob(x = NULL, out.path = NULL)
}
\arguments{
\item{x}{`data.frame` or `NULL`.
Input dataset. Must contain all required columns defined in
`required.inputs`. If `NULL`, an interactive file chooser is launched.}

\item{out.path}{`character` or `NULL`.
File path for the output `.txt` file. If `NULL`, a path is generated
automatically.}
}
\description{
`skeleton_EMOb()` processes BRUV/DOV/ROV analysis datasheets and exports a 
structured EMOb skeleton `.txt` file for use in EventMeasure. The function writes 
a formatted EMOb skeleton file to disk and no object is returned to the R 
environment. The function is designed for both scripted and interactive use,
 depending on whether the input `x` is supplied.
}
\details{
When `x` is provided, it must be a `data.frame`.  If `x` is `NULL`, the 
user is prompted to choose an input dataset to use as `x`. This should be 
an analysis datasheet within a campaign on iLab_Fish. 

The columns in `x` are validated to confirm that all columns listed in the 
internal vector `required.inputs` are present. If any required column is
missing, the function stops and informs the user which columns are missing.
}
\examples{
\dontrun{
# Fully interactive use:
skeleton_EMOb()

# Scripted use with a data frame:
skeleton_EMOb(
  x = example_df,
  out.path = "outputs/emob_skeleton.txt"
)
}

}
