% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_metadata_functions.R
\name{add_missing_variables}
\alias{add_missing_variables}
\title{Add missing variables to field metadata}
\usage{
add_missing_variables(x, missing)
}
\arguments{
\item{x}{A data frame containing field metadata.}

\item{missing}{A vector containing the names of expected variables not found
field metadata}
}
\description{
Add blank columns for non-vital variables missing from field metadata.
}
\examples{
\dontrun{
field.data <- add_missing_variables(field.data, missing)
}

}
