% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_ilab_structure.R
\name{campaign_structure}
\alias{campaign_structure}
\title{Build the default directory/file specification}
\usage{
campaign_structure(MP = NULL, campaign = NULL)
}
\arguments{
\item{MP}{(Optional) acronym(s) of target marine park(s).}

\item{campaign}{(Optional) name of campaign(s) in the target marine park(s). 
Requires `MP` to be defined.}
}
\value{
A data frame with the required columns for `check_structure()`:
  \describe{
    \item{target}{Describing the target directories/files for that row.}
    \item{type}{`"file"` or `"dir"`.}
    \item{rel_dir_pattern}{Regex matching the relative parent directory.}
    \item{name_regex}{Regex matching the basename (file or folder name).}
    \item{required}{Logical, whether at least `min` matches are required.}
    \item{min}{Minimum number of matches (per rule).}
    \item{max}{Maximum allowed matches (use `Inf` for unlimited).}
    \item{case_sensitive}{Logical, case sensitivity for file/dir names.}
    \item{rel_dir_case_sensitive}{Logical, case sensitivity for folder matching.}
    }
    
    @importFrom tibble tribble
}
\description{
Internal function used in `check_structure()` to create a data frame describing
the expected file/folder format used in iLab Fish campaigns. The output data frame
includeds regex patterns used to search for directory and file combinations.
}
