% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dir.R
\name{get_dir}
\alias{get_dir}
\title{Get a OneDrive directory path}
\usage{
get_dir(
  dir,
  sub.folder = NULL,
  root = Sys.getenv("OneDriveCommercial"),
  exclude = c("Apps", "Attachments", "Desktop", "Documents", "Pictures",
    "OneNote Loop Files", "Microsoft Copilot Chat Files", "Microsoft Teams Chat Files")
)
}
\arguments{
\item{dir}{Character vector for a top-level folder under \code{root} (e.g.
"ilab-fish"). Only requires a partial match for a folder and errors if multiple 
matches are found.}

\item{sub.folder}{Optional character vector of a subdirectory within \code{dir} (e.g.
"NCMP/2024-02_BRUV").}

\item{root}{Character vector setting the root for \code{dir}. Defaults to
\code{Sys.getenv("OneDriveCommercial")} which should be SharePoint for DBCA
staff who sync sharepoint to their PC. You can override (e.g.,\code{Sys.getenv("OneDrive")}).}

\item{exclude}{Character vector containing folder names to exclude from the available top-level
folders in \code{root}. Defaults to common system/personal folders.}
}
\value{
A length-1 character vector giving the normalized file path.
}
\description{
Validates and returns the full path to a first-level folder (and optional
subfolder) under a configurable root directory (defaults to OneDrive Buisness).
This allows DBCA personnel to create shareable scripts that can access OneDrive 
files synced to their PC.
}
\examples{
\dontrun{
# Exact match to a OneDrive Business top-level folder:
get_dir("ilab_fish")

# With an existing subfolder:
get_dir("ilab_fish", sub.folder = "NCMP/2024-02_BRUV")

# Use a different root (e.g., personal OneDrive):
get_dir("Photos", root = Sys.getenv("OneDrive"))
}

}
\seealso{
\link[base]{list.dirs}, \link[base]{file.path}, \link[base]{Sys.getenv}
}
